fix(cdc_acm): Fix Receive buffer append function on ESP32-P4
The Receive buffer append function changes location of the data buffer.
While this works fine on S2 and S3, it will crash on P4 because now we must do
cache sync for the data buffer. This will crash on unaligned cache sync,
so this feature is disabled on P4 for now.